+++ CHANGELOG.rst
@@ -32,6 +32,8 @@ Added
   keys (`#571 <https://github.com/omni-us/jsonargparse/pull/571>`__).
 - ``dataclass`` types now accept ``class_path`` and ``init_args`` as value
   (`#581 <https://github.com/omni-us/jsonargparse/pull/581>`__).
+- Support for PEP 692 (i.e., ``Unpack[TypedDict]`` annotations for ``**kwargs``)
+  (`#586 <https://github.com/omni-us/jsonargparse/pull/586>`)
 
 Fixed
 ^^^^^